+++ src/request.rs
@@ -1,4 +1,3 @@
-use colored::*;
 use reqwest::blocking::Client;
 use reqwest::header::AUTHORIZATION;
 use reqwest::header::CONTENT_TYPE;
@@ -12,8 +11,7 @@ use crate::Arguments;
 use crate::{MODEL_DEFAULT, NUMBER_DEFAULT, TEMPERATURE_DEFAULT, TOP_P_DEFAULT};
 
 // OPENAI URLS
-const COMPLETIONS_URL: &str = "/v1/completions";
-const MODELS_URL: &str = "/v1/models";
+const COMPLETIONS_URL: &str = "/v1/chat/completions";
 
 const MAX_TOKENS: u32 = 1000;
 
@@ -32,9 +30,16 @@ struct Usage {
 #[derive(Deserialize)]
 #[allow(dead_code)]
 struct Choice {
-    text: String,
+    message: Message,
     index: u16,
     finish_reason: String,
+}
+
+#[derive(Deserialize)]
+#[allow(dead_code)]
+struct Message {
+    role: String,
+    content: String,
 }
 #[derive(Deserialize)]
 struct Response {
@@ -56,16 +61,6 @@ struct Version {
     num: String,
 }
 
-#[derive(Deserialize)]
-struct Models {
-    data: Vec<Model>,
-}
-
-#[derive(Deserialize)]
-struct Model {
-    id: String,
-}
-
 /// Get completions from input prompt
 pub fn completions(arguments: Arguments, config: Config) -> Result<String, String> {
     let number = arguments.number.unwrap_or(NUMBER_DEFAULT);
@@ -77,34 +72,25 @@ pub fn completions(arguments: Arguments, config: Config) -> Result<String, Strin
         .unwrap_or_else(|| config.model.unwrap_or_else(|| String::from(MODEL_DEFAULT)));
     let body = json!({ 
         "model": model, 
-        "prompt": arguments.prompt.unwrap_or_default(),  
-        "max_tokens": MAX_TOKENS, 
+        "max_tokens": MAX_TOKENS,
+        "messages": [{"role": "user", "content": arguments.prompt.unwrap_or_default()}],
         "n": number, 
         "temperature": temperature, 
-        "top_p": top_p,
-    "echo": arguments.echo });
+        "top_p": top_p});
 
     let json_response = post_openai(config.token, COMPLETIONS_URL.to_string(), body)?;
-    let response: Response =
-        serde_json::from_str(&json_response).or(Err("Could not serialize to CargoResponse"))?;
+    let response: Response = serde_json::from_str(&json_response)
+        .or(Err("Could not serialize response from chat completion"))?;
 
     let output = response
         .choices
         .into_iter()
-        .map(|x| x.text)
+        .map(|x| x.message.content)
         .collect::<Vec<String>>()
         .join("\n\n---");
     let suffix = arguments.suffix.unwrap_or_default();
 
     Ok(format!("{output}{suffix}"))
-}
-pub fn models(config: Config) -> Result<String, String> {
-    let json_response = get_openai(config.token, MODELS_URL.to_string())?;
-    let models: Models =
-        serde_json::from_str(&json_response).or(Err("Could not serialize to Models"))?;
-    let model_list: Vec<String> = models.data.into_iter().map(|model| model.id).collect();
-    let result = format!("{}\n\n{}", "Models: ".green(), model_list.join("\n"));
-    Ok(result)
 }
 
 fn post_openai(token: String, url: String, body: serde_json::Value) -> Result<String, String> {
@@ -119,28 +105,6 @@ fn post_openai(token: String, url: String, body: serde_json::Value) -> Result<St
         .header(CONTENT_TYPE, "application/json")
         .header(AUTHORIZATION, authorization)
         .json(&body)
-        .send()
-        .or(Err("Did not get response from server"))?;
-    sp.stop();
-    print!("\x1b[2K\r");
-
-    if response.status().is_success() {
-        Ok((response.text()).or(Err("Could not read response text"))?)
-    } else {
-        Err(format!("Error: {:#?}", response.text()))
-    }
-}
-
-fn get_openai(token: String, url: String) -> Result<String, String> {
-    let openai_url: &str = "https://api.openai.com";
-    let request_url = format!("{openai_url}{url}");
-    let authorization: &str = &format!("Bearer {token}");
-
-    let mut sp = Spinner::new(SPINNER, MESSAGE.into());
-    let response = Client::new()
-        .get(request_url)
-        .header(CONTENT_TYPE, "application/json")
-        .header(AUTHORIZATION, authorization)
         .send()
         .or(Err("Did not get response from server"))?;
     sp.stop();