+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+- Set the default model to `gpt-3.5-turbo`
+- Use chat completions endpoint (so that GPT-4 can be used)
+- Remove the models flag as it no longer applies for the chat completions endpoint
+- Remove the echo flag as it is no longer supported by the new endpoint
+
 ## 2023-03-19 0.1.5
 
-- Remove spinner once it stops
+- Remove the spinner once it stops
 
 ## 2023-03-19 0.1.4